fixed the build-resources.sh script.  was using -RfT, should have been -Rf

+++ teardown.sh
@@ -2,9 +2,10 @@
 if [ -z $srvAcct ]; then
 	srvAcct="fivem"
 fi
-
 su $srvAcct -c "screen -XS 'fivem' quit"
-rm -rf /home/$srvAcct
-rm -rf /var/software
 deluser $srvAcct
+if [ -d "/home/$srvAcct" ]; then
+	rm -rf /home/$srvAcct
+fi
+rm -rf /var/software
 mysql -e "DROP DATABASE essentialmode;"
